# swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8889'
basePath: /
...
paths:
...
  /posts:
    post:
      tags:
        - stwitter-controller
      summary: createPost
      operationId: createPostUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: switterViewModel
          description: switterViewModel
          required: true
          schema:
            $ref: '#/definitions/SwitterViewModel'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/SwitterViewModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/posts/user/{poster-name}':
    get:
      tags:
        - stwitter-controller
      summary: getPostByPosterName
      operationId: getPostByPosterNameUsingGET
      produces:
        - '*/*'
      parameters:
        - name: posterName
          in: path
          description: posterName
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/SwitterViewModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/posts/{id}':
    get:
      tags:
        - stwitter-controller
      summary: getPost
      operationId: getPostUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SwitterViewModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  Comment:
    type: object
    properties:
      commentContent:
        type: string
      commentId:
        type: integer
        format: int32
      commenterName:
        type: string
      createDate:
        type: string
        format: date
      postId:
        type: integer
        format: int32
    title: Comment
...
  SwitterViewModel:
    type: object
    properties:
      comments:
        type: array
        items:
          $ref: '#/definitions/Comment'
      id:
        type: integer
        format: int32
      postContent:
        type: string
      postDate:
        type: string
        format: date
      posterName:
        type: string
    title: SwitterViewModel
...